fixed truncated long
Former-commit-id: bae1d17e1b6ee1328d2f6c1b97a528c8a23c8e44

+++ console/src/main/java/org/janelia/it/FlyWorkstation/shared/util/Utils.java
@@ -611,7 +611,7 @@ public class Utils {
             output.write(buffer, 0, n);
             count += n;
             if (worker!=null) {
-                worker.setProgress(count, (int)length);
+                worker.setProgress(count, length);
                 if (worker.isCancelled()) throw new CancellationException();
             }
         }
